+++ models/user.js
@@ -1,22 +1,14 @@
 'use strict';
 
-module.exports = function (sequelize, DataTypes) {
-  var User = sequelize.define(
-    'User',
-    {
-      name: DataTypes.STRING,
-      nickname: DataTypes.STRING,
-      provider: DataTypes.ENUM('Local', 'Facebook'),
-      loginId: DataTypes.STRING,
-      photo: DataTypes.STRING
-    },
-    {
-      classMethods: {
-        associate: function (models) {
-          models.User.hasMany(models.Post);
-        }
-      }
-    });
+var mongoose = require( 'mongoose' );
+var Schema   = mongoose.Schema;
 
-  return User;
-};
+var userSchema = new Schema({
+  name: String,
+  nickname: String,
+  provider: String,
+  loginId: String,
+  photo:String
+});
+
+module.exports = mongoose.model('User', userSchema);
